SAK-7399
git-svn-id: https://source.sakaiproject.org/svn/sam/branches/sakai_2-3-x@19670 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ samigo-services/src/java/org/sakaiproject/tool/assessment/facade/PublishedAssessmentFacadeQueries.java
@@ -1853,7 +1853,7 @@ public class PublishedAssessmentFacadeQueries
 	    final HibernateCallback hcb = new HibernateCallback(){
 		    public Object doInHibernate(Session session) throws HibernateException, SQLException {
 			Query q = session.createQuery(hql);
-			q.setLong(0, publishedId);
+			q.setLong(0, publishedId.longValue());
 			return q.list();
 		    };
 		};
